Time: 0 ms (100%), Space: 21.8 MB (95.03%) - LeetHub

+++ 0098-validate-binary-search-tree/0098-validate-binary-search-tree.cpp
@@ -11,15 +11,15 @@
  */
 class Solution {
 public:
-    bool solve(TreeNode* root, int mini, int maxi){
+    bool solve(TreeNode* root, long long mini, long long maxi){
         if(root == nullptr) return true;
         if(root->val > mini && root->val < maxi)
         return solve(root->left, mini, root->val) && solve(root->right, root->val, maxi);
         else return false;
     }
     bool isValidBST(TreeNode* root) {
-        int mini = -1000000000000;
-        int maxi = 1000000000000;
+        long long mini = -1000000000000;
+        long long maxi = 1000000000000;
         return solve(root, mini, maxi);
     }
 };
